[LandingPage] do not wrap landing page with AppLayout
