+++ components/AppLayout/index.js
@@ -1,2 +1,5 @@
 import AppLayout from './AppLayout';
+export { default as LoginModal } from './LoginModal';
+export { default as AppFooter } from './AppFooter';
+
 export default AppLayout;
